% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emorph2.R
\docType{data}
\name{Canidae}
\alias{Canidae}
\title{Canidae dataset}
\format{
Four objects of different classes
}
\usage{
data(Canidae)
}
\description{
Canidae dataset
}
\details{
\itemize{
\item W Pooled within variance covariance matrix. Can be used as the G matrix in examples.
\item means Matrix containing mean trait values for species. Total of 39 species (lines) and 40 traits (columns).
\item n.s Table containing the sample size for each species. 39 values.
\item tree Dated phylogenetic tree representing the relationship among taxa and node age.
}
}
\keyword{datasets}
