% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdists.R
\name{cdists}
\alias{cdists}
\title{Calculate interlandmark distances}
\usage{
cdists(X, dists, names = rownames(X))
}
\arguments{
\item{X}{Three dimentional array of form data.}

\item{dists}{Distances to be extracted}

\item{Names}{Landmark names}
}
\value{
A dataframe of interlandmark distances per specimen.
}
\description{
xxxxxxxx
}
\details{
xxxxxxxxx
}
\seealso{
\code{\link[ape]{pic}}
}
\author{
Fabio Andrade Machado
}
