% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTransform.R
\name{covProd}
\alias{covProd}
\title{Covariance of a random variable and the product of two random variables}
\usage{
covProd(V, m)
}
\arguments{
\item{V}{Covariance matrix of three or four traits.}

\item{m}{Vector of expectations of the variables being multiplied}
}
\description{
Covariance of a random variable and the product of two random variables
}
\author{
Fabio Andrade Machado
}
