% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTransform.R
\name{covScalar}
\alias{covScalar}
\title{Covariance of the product a random variable by a scalar}
\usage{
covScalar(cov, s1 = 1, s2 = 1)
}
\arguments{
\item{cov}{Covariance between two random variables}

\item{s1}{multiplying scalar for the first variable}

\item{s2}{multiplying scalar for the second variable}
}
\description{
Covariance of the product a random variable by a scalar
}
\author{
Fabio Andrade Machado
}
