% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTransform.R
\name{covSum}
\alias{covSum}
\title{Covariance of a random variable and the sum of two random variables}
\usage{
covSum(V, covxy = NULL, covxu = NULL)
}
\arguments{
\item{V}{Covariance matrix of three traits, the second and third traits are the ones being summed.}

\item{covxy}{Covariance between two random variables}

\item{covxu}{Covariance between two random variables}
}
\description{
Covariance of a random variable and the sum of two random variables
}
\author{
Fabio Andrade Machado
}
