% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenQst.R
\name{eigenQst}
\alias{eigenQst}
\title{A multivariate Qst analysis}
\usage{
eigenQst(B, G)
}
\arguments{
\item{B}{A covariance matrix between population means.}

\item{G}{A genetic covariance matrix.}
}
\description{
A multivariate Qst analysis
}
\author{
Fabio Andrade Machado
}
