% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTransform.R
\name{expRat}
\alias{expRat}
\title{Expectation of the ratio of two random variables}
\usage{
expRat(V, m)
}
\arguments{
\item{V}{Covariance matrix of two traits}

\item{m}{Vector of trait means}
}
\description{
Expectation of the ratio of two random variables
}
\author{
Fabio Andrade Machado
}
