% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNfuseReps.R
\name{fixNfuseReps}
\alias{fixNfuseReps}
\title{Fix and fuse replicates}
\usage{
fixNfuseReps(rep1, rep2, pairedLM)
}
\arguments{
\item{pairdLM}{}
}
\description{
Fix and fuse replicates
}
\author{
Fabio Andrade Machado
}
