% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landeDist.R
\name{landeDist}
\alias{landeDist}
\title{Lande's generalized genetic distance}
\usage{
landeDist(dz, G, Ne)
}
\description{
Calculates Lande's generalized genetic distance
}
\references{
Lande, R. 1979. Quantitative genetic analysis of multivariate evolution, applied to brain: body size allometry. Evolution, 33(1), 402–416.
}
\author{
Fabio Andrade Machado
}
