% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landeDist.R
\name{landeDist.test}
\alias{landeDist.test}
\title{Lande's generalized genetic distance}
\usage{
landeDist.test(
  dz,
  G,
  Ne,
  CI = 0.95,
  par = FALSE,
  nsim = 999,
  alpha = 0.05,
  MonteCarlo = FALSE,
  n,
  parallel = FALSE
)
}
\description{
Test for the presence of preferencial lines of evolution among multiple
independent lineages/replicates.
}
\references{
Lande, R. 1979. Quantitative genetic analysis of multivariate evolution, applied to brain: body size allometry. Evolution, 33(1), 402–416.
}
\author{
Fabio Andrade Machado
}
