% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeShift.R
\name{node_shift}
\alias{node_shift}
\title{Shifting nodes}
\usage{
node_shift(phy, criteria, shift = NULL)
}
\arguments{
\item{phy}{an object of class "phylo"}

\item{criteria}{icriteri}
}
\value{
A
}
\description{
rescaling phy
}
\details{
S
}
\seealso{
\code{\link[ape]{pic}}
}
\author{
Fabio Andrade Machado
}
