% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npEigentest.R
\name{npEigentest}
\alias{npEigentest}
\title{npEigentest}
\usage{
npEigentest(G, means, phy, n.s, sims = 1000, dim.ret = NULL,
  parallel = FALSE)
}
\arguments{
\item{G}{matrix kxk for k number of traits.}

\item{phy}{a phylogenetic tree. Must be of class 'phylo'.}

\item{n.s}{a vector indicating the sample sizes for each terminal.}

\item{dim.ret}{choose the number of dimentions that should be retained in the
analysis.}

\item{parallel}{Should parallelize? Default is FALSE. See 'parallel vignette'
for details}
}
\description{
npEigentest produces non-parametric confidence intervals spectral
decomposition-based tests of matrix proportionality.
}
\examples{
data("Canidae_means")
data("Canidae_tree")
data("Canidae_sample_sizes")
data("Canidae_W")
test.out<-npEigentest(G=W, means, tree, n.s, sims = 100, dim.ret = 20)
test.out$SimValues
}
