% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multiphylo.R
\name{sim_multiphylo}
\alias{sim_multiphylo}
\title{Quantitative genetics-based simulation of multivariate evolution.}
\usage{
sim_multiphylo(
  G,
  phy,
  n.s,
  selection = "random",
  efsize = 0,
  gen_time = 1e-06,
  Nef = 1000,
  Nef_osc = "no",
  matrix = TRUE,
  Nef_par = NULL,
  scale = FALSE,
  means = NULL
)
}
\arguments{
\item{G}{matrix kxk for k number of traits.}

\item{phy}{a phylogenetic tree. Must be of class 'phylo'.}

\item{n.s}{a vector indicating the sample sizes for each terminal.}

\item{selection}{define the type oif selection "random" or numeric between 1
and k.}

\item{efsize}{strenght of selection relative to drif. Must be \code{>= 0}.}

\item{gen_time}{generation time in the same time unity as the phylogenetic
tree (usually MY).}

\item{Nef}{effective population size for initial population. Must be >= 1.}

\item{Nef_osc}{oscilation in effective population size (Nef) over time.
Character "no", "norm", or "unif".}

\item{matrix}{logic TRUE of FALSE. Should the output be a set
of marices? See 'Value' for more details.}

\item{Nef_par}{set parameters for the oscilation of effective population
size (Nef).}

\item{scale}{logic TRUE of FALSE. Should the matrices be scaled to the
empirical character values (means)?}

\item{means}{matrix sxk containing the empirical means of all k characters
for each species (s). Default = NULL.}
}
\value{
The 'sim_multiphylo' function returns an object of class "list". This is a
list of items of class "matrix", all kxk.
If 'matrix' = TRUE, the function returns 5 variance covariance matrices:
\describe{
  \item{R}{Evolutionary rate matrix. Describes the rate of evolution and
  co-evolution among characters.
  Diagonal contains traits' rate evolution according to Brownian-Motion,
  off diagonals represent traits co-evolution.}
  \item{W}{Within matrix. Pooled-within species' trait variance-covariance matrix.}
  \item{B}{Between matrix. Variance-vovariance between average species traits.}
  \item{G}{Genetic matrix. The original aditive variance-covariance G matrix imputed.}
  \item{A}{Selection matrix. The expected effect due to selection.}
 }
If 'matrix' = FALSE, the function returns 3 matrices:
 \describe{
  \item{bdata}{Simulated trait values for species averages.}
  \item{wdata}{Simulated intraspecific error.}
  \item{G}{the original aditive variance-covariance G matrix imputed.}
  }
}
\description{
Simulates evolution using Lande's multivariate breeder's equation.
}
