% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDrift.R
\name{simDrift}
\alias{simDrift}
\title{Simulation of drift in a metapopulation..}
\usage{
simDrift(G, theta, nsims = 1, mu = matrix(0, nrow(theta), nrow(G)))
}
\arguments{
\item{G}{matrix kxk for k number of traits.}

\item{theta}{matrix of among-population coancestry coefficients.}
}
\value{
A nxkxnsims cubic array of simulated population averages.
}
\description{
Simulates evolution by genetic drift on a metapopulation with varying levels
of geneflow.
}
