% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTransform.R
\name{varProd}
\alias{varProd}
\title{Variance of the product of two random variables}
\usage{
varProd(V, m)
}
\arguments{
\item{V}{Covariance matrix of two traits}

\item{m}{Vector of trait means}
}
\description{
Variance of the product of two random variables
}
\author{
Fabio Andrade Machado
}
