% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTransform.R
\name{varRat}
\alias{varRat}
\title{Variance of the ratio of two random variables}
\usage{
varRat(V, m)
}
\arguments{
\item{V}{Covariance matrix of two traits}

\item{m}{Vector of trait means}
}
\description{
Variance of the ratio of two random variables
}
\author{
Fabio Andrade Machado
}
