% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTransform.R
\name{varScalar}
\alias{varScalar}
\title{Variance of the product a random variable by a scalar}
\usage{
varScalar(var, s)
}
\arguments{
\item{var}{Variance of a random variable}

\item{s}{multiplying scalar}
}
\description{
Variance of the product a random variable by a scalar
}
\author{
Fabio Andrade Machado
}
