% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTransform.R
\name{varSum}
\alias{varSum}
\title{Variance of the sum of two random variables}
\usage{
varSum(V)
}
\arguments{
\item{V}{Covariance matrix of two traits}
}
\description{
Variance of the sum of two random variables
}
\author{
Fabio Andrade Machado
}
