% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftsel.R
\name{driftsel}
\alias{driftsel}
\title{Multivariate test for between-population selection}
\usage{
driftsel(G, means, theta, anc = NULL, sims = 0, parallel = FALSE)
}
\arguments{
\item{G}{Genetic additive among-trait covariance matrix for k number of
traits. Can be either a single kxk matrix, a list of multiple Gs or
three-dimensional array (kxkxi), with i being the number of items.}

\item{means}{matrix containing the means/fixed effects of all k characters
for each population (n). Can be either a single nxk matrix, a list of
multiple nxk matrices or a three-dimensional array (nxkxi). Must match
the dimensionality of G.}

\item{theta}{The coancestry coefficient among populations. Can be either a
single nxn matrix, a list of nxn matrices of a three-dimensional array
(nxnxi).}

\item{anc}{matrix containing the ancestral state of all k characters. Can
be either a single nxk matrix, a list of multiple nxk matrices or a
three-dimensional array (nxkxi). Must match the dimensionality of G. If not
provided, the ancestral state is assumed to be 0 for all traits.}

\item{sims}{Numeric. Number of simulations used for the non-parametric tests.
If NULL, the parametric test is performed.}

\item{parallel}{Should be parallelized? Default is FALSE.}

\item{pop.test}{logical. If TRUE, calculate the mahalanobis distance for each
population based on the expected divergence.
simulations.}
}
\value{
If nsim=0 returns a single value for the parametric S-test. If nsim>0 returns
a non-parametric S-test value for each population.
}
\description{
Tests for the presence of directional and stabilizing selection on
multivariate traits for a multi-population dataset.
}
\author{
Fabio Andrade Machado
}
