% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenPar.R
\name{eigenPar}
\alias{eigenPar}
\title{A multivariate test of parallel evolution between multiple lineages}
\usage{
eigenPar(
  X1,
  X2 = NULL,
  null = c("random", "drift"),
  sims = 999,
  alpha = c(0.025, 0.975),
  G = NULL,
  parallel = FALSE
)
}
\arguments{
\item{X1}{A matrix of multivariate averages of lineages before selection/
experimental manipulation OR a multivariate matrix of differences between
before and after selection.}

\item{X2}{NULL or a matrix of multivariate averages of lineages before
selection. If a value is provided, it has to be of the same dimentions as X1}

\item{null}{Type of null hypothesis to be tested. If null="random", tests if
empirical vectors differs from totally random directions. If null="drift",
function tests if vectors differsn from what is expected under drift. A G
matrix has to be provided}

\item{G}{Default to NULL. A matrix of trait-by-trait genetic covariances.
Only used if null="drift".}
}
\description{
Test for the presence of preferencial lines of evolution among multiple
independent lineages/replicates.
}
\references{
De Lisle, S. P., & Bolnick, D. I. 2020. A multivariate view of parallel evolution. Evolution, 65, 1143–16. http://doi.org/10.1111/evo.14035
}
\seealso{
\code{\link[Morpho]{angleTest}}
}
\author{
Fabio Andrade Machado
}
