% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.R
\name{psc}
\alias{psc}
\title{Phylogenetic serial contrasts}
\usage{
psc(x, phy)
}
\arguments{
\item{x}{a numeric vector.}

\item{phy}{an object of class "phylo"}
}
\value{
what to return?
}
\description{
Calculate McPeek's phylogenetic contrasts along all banches of a phylogeny
}
\details{
Some details here
}
